+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Xmutarn v0.27.0 (2019-xx-xx) - Release 10
 
+- Adicionado classe Typescipt responsável por gerenciar uma barra de ferramentas.
 - Adicionado ícones para toasts. Issue: [#9](https://github.com/vinivsl/Xmutarn/issues/9).
 - Adicionado a exibição e omissão de overlay para toasts, quando o **timeout** for setado como **0**.